requirements.txt
streamlit>=1.34.0
crewai>=0.121.0
python-dotenv>=1.0.1